+++ src/modules/blog/article-category/article-category.module.ts
@@ -1,0 +1,16 @@
+import { Module } from '@nestjs/common';
+import { ArticleCategoryService } from './article-category.service';
+import { ArticleCategoryController } from './article-category.controller';
+import { ArticleCategory, ArticleCategorySchema } from 'src/schemas/blog/article-category.schema';
+import { MongooseModule } from '@nestjs/mongoose';
+import { JwtStrategy } from 'src/jwt/jwt.strategy';
+
+const ARTICLE_CATEGORY_MONGO_MODULE = MongooseModule.forFeature([{ name: ArticleCategory.name, schema: ArticleCategorySchema }], 'blog');
+
+@Module({
+  imports: [ARTICLE_CATEGORY_MONGO_MODULE],
+  controllers: [ArticleCategoryController],
+  providers: [ArticleCategoryService, JwtStrategy],
+  exports: [ArticleCategoryService],
+})
+export class ArticleCategoryModule {}