+++ src/modules/blog/blog.module.ts
@@ -4,13 +4,15 @@ import { BlogController } from './blog.controller';
 import { MongooseModule } from '@nestjs/mongoose';
 import { RouterModule } from '@nestjs/core';
 import { ArticleModule } from './article/article.module';
+import { ArticleCategoryModule } from './article-category/article-category.module';
 
 @Module({
   imports: [
-    MongooseModule.forRoot('mongodb://yrr:AlyYrrAdmin123@120.79.162.189:5606/money?authSource=admin', {
+    MongooseModule.forRoot('mongodb://yrr:AlyYrrAdmin123@120.79.162.189:5606/blog?authSource=admin', {
       connectionName: 'blog',
     }),
     ArticleModule,
+    ArticleCategoryModule,
     RouterModule.register([
       {
         path: 'blog',
@@ -18,6 +20,10 @@ import { ArticleModule } from './article/article.module';
           {
             path: '/',
             module: ArticleModule,
+          },
+          {
+            path: '/',
+            module: ArticleCategoryModule,
           },
         ],
       },