+++ CHANGELOG.md
@@ -5,9 +5,13 @@
 - 添加blog的article和article-category的schemas目录
 - 创建文章article 数据库集合
 - 用户昵称字段name改成nickName
-##### 文章接口
+- 创建文章article-category 数据库集合
+##### 文章
 - 条件并分页获取文章列表
 - 新增文章
+##### 文章分类
+- 新增文章分类
+- 获取全部文章分类列表
 
 ## 1.1.0__02
 - 修改登录日志